Build Rating

+++ app/Http/Controllers/Api/Apartments/ApartmentController.php
@@ -30,7 +30,8 @@ class ApartmentController extends Controller
 
     public function show($id)
     {
-        $apartment = Apartment::with('images', 'owner')->findOrFail($id);
+        $apartment = Apartment::where('is_available', true)
+        ->with('images', 'owner')->findOrFail($id);
 
         return response()->json([$apartment, 200]);
     }
@@ -132,14 +133,20 @@ public function update(Request $request, $id)
         
         
         // To Delete Existing Photos
-        if ($request->filled('delete_ids')) {
-            $imagesToDelete = $apartment->images()->whereIn('id', $request->delete_ids)->get();
+       if ($request->filled('delete_ids')) {
+        $imagesToDelete = $apartment->images()->whereIn('id', $request->delete_ids)->get();
 
-            foreach ($imagesToDelete as $image) {
+        foreach ($imagesToDelete as $image) {
+            
+        
+            if (Storage::disk('public')->exists($image->image_path)) {
                 Storage::disk('public')->delete($image->image_path);
-                $image->delete();
             }
+            
+            
+            $image->delete(); 
         }
+    }
 
 
         // To Add New Photos
@@ -175,19 +182,32 @@ public function update(Request $request, $id)
 
     // delete apartment
     /////////////////////////////////////////////////////////////////////////////////////////////////
-    public function destroy($id)
-    {
-        $apartment = Apartment::findOrFail($id);
-        if ($apartment->owner_id !== Auth::id()) {
-            return response()->json([
-                'message' => 'Unauthorized action. You do not own this apartment.',
-            ], 403);
-        }
-        $apartment->delete();
-
-        return response()->json(['The Apartment Deleted Successfully', 200]);
+  public function destroy($id)
+{
+    $apartment = Apartment::findOrFail($id);
+    
+    if ($apartment->owner_id !== Auth::id()) {
+        return response()->json([
+            'message' => 'Unauthorized action. You do not own this apartment.',
+        ], 403);
     }
+    
+    $imagesToDelete = $apartment->images()->get(); 
+    
+    foreach ($imagesToDelete as $image) {
+        
+        if (Storage::disk('public')->exists($image->image_path)) {
+            Storage::disk('public')->delete($image->image_path);
+        }
+        
+        $image->delete(); 
+    }
+    
+    
+    $apartment->delete(); 
 
+    return response()->json(['message' => 'The Apartment Deleted Successfully'], 200);
+}
     // fillter
     /////////////////////////////////////////////////////////////////////////////////////////////////
